app/requirements.txt
 Flask==2.3.3 
 pytest==7.4.2
 python-geoip-python3==1.3
 